ghidra_azure/decompile_handler/requirements.txt
requests==2.31.0
azure-identity==1.14.0
azure-mgmt-containerinstance==10.1.0
azure-storage-blob==12.16.0
